+++ src/Core/Baskets/Services/BasketService.php
@@ -13,6 +13,7 @@ use GetCandy\Api\Core\Baskets\Models\SavedBasket;
 use GetCandy\Api\Core\Baskets\Events\BasketStoredEvent;
 use GetCandy\Api\Core\Baskets\Interfaces\BasketFactoryInterface;
 use GetCandy\Api\Core\Products\Interfaces\ProductVariantInterface;
+use GetCandy\Api\Core\Currencies\Interfaces\CurrencyConverterInterface;
 
 class BasketService extends BaseService
 {
@@ -29,6 +30,13 @@ class BasketService extends BaseService
     protected $factory;
 
     /**
+     * The currency converter
+     *
+     * @var CurrencyConverterInterface
+     */
+    protected $currencies;
+
+    /**
      * The variant factory.
      *
      * @var ProductVariantInterface
@@ -37,11 +45,13 @@ class BasketService extends BaseService
 
     public function __construct(
         BasketFactoryInterface $factory,
-        ProductVariantInterface $variantFactory
+        ProductVariantInterface $variantFactory,
+        CurrencyConverterInterface $currencies
     ) {
         $this->model = new Basket();
         $this->factory = $factory;
         $this->variantFactory = $variantFactory;
+        $this->currencies = $currencies;
     }
 
     /**
@@ -65,9 +75,7 @@ class BasketService extends BaseService
             $basket->user()->associate($user);
         }
 
-        if (! $basket->currency) {
-            $basket->currency = app('api')->currencies()->getDefaultRecord()->code;
-        }
+        $basket->currency = $this->currencies->get()->code;
 
         $basket->save();
 
@@ -167,6 +175,7 @@ class BasketService extends BaseService
      * Store a basket.
      *
      * @param array $data
+     * @param ?User $user
      *
      * @return Basket
      */
@@ -177,16 +186,61 @@ class BasketService extends BaseService
             $user
         );
 
-        $basket->meta = $data['meta'] ?? null;
-
-        if (empty($data['currency'])) {
-            $basket->currency = app('api')->currencies()->getDefaultRecord()->code;
-        } else {
-            $basket->currency = $data['currency'];
-        }
+        $basket = $this->setupBasket($basket, $data);
 
         $basket->lines()->delete();
 
+        return $this->storeAndUpdateBasket($basket, $data);
+    }
+
+    /**
+     * Add new lines to a basket, without remapping the existing lines.
+     *
+     * @param array $data
+     * @param ?User $user
+     *
+     * @return Basket
+     */
+    public function addLines(array $data, $user = null)
+    {
+        $basket = $this->getBasket(
+            ! empty($data['basket_id']) ? $data['basket_id'] : null,
+            $user
+        );
+
+        $this->setupBasket($basket, $data);
+
+        return $this->storeAndUpdateBasket($basket, $data);
+    }
+
+    /**
+     * @param Basket $basket
+     * @param array $data
+     * @return Basket
+     */
+    protected function setupBasket(Basket $basket, array $data)
+    {
+        if (isset($data['meta'])) {
+            $basket->meta = $data['meta'];
+        }
+
+        if (isset($data['currency'])) {
+            $basket->currency = $data['currency'];
+        }
+        if (is_null($basket->currency)) {
+            $basket->currency = app('api')->currencies()->getDefaultRecord()->code;
+        }
+
+        return $basket;
+    }
+
+    /**
+     * @param Basket $basket
+     * @param array  $data
+     * @return Basket
+     */
+    protected function storeAndUpdateBasket(Basket $basket, array $data)
+    {
         if (! empty($data['variants'])) {
             $this->remapLines($basket, $data['variants']);
         }
@@ -260,12 +314,21 @@ class BasketService extends BaseService
     {
         $service = app('api')->productVariants();
 
-        $variants = collect($variants)->map(function ($item) use ($service, $basket) {
-            $variant = $this->variantFactory->init(
-                $service->getByHashedId($item['id'])
-            )->get($item['quantity']);
+        $collectedVariants = [];
 
-            return [
+        // Collect variants with the same ID and add up their quantities
+        collect($variants)->each(function ($item) use (&$collectedVariants, $service) {
+            $variant = $this->variantFactory
+                ->init($service->getByHashedId($item['id']))
+                ->get($item['quantity']);
+
+            if (array_key_exists($variant->id, $collectedVariants)) {
+                $collectedVariants[$variant->id]['quantity'] += $item['quantity'];
+
+                return;
+            }
+
+            $collectedVariants[$variant->id] = [
                 'product_variant_id' => $variant->id,
                 'quantity' => $item['quantity'],
                 'total' => $item['quantity'] * $variant->price,
@@ -273,7 +336,19 @@ class BasketService extends BaseService
             ];
         });
 
-        $basket->lines()->createMany($variants->toArray());
+        // If a basket line with this variant already exists, increase that instead
+        $basket->lines->map(function ($line) use (&$collectedVariants) {
+            $variant_id = $line->product_variant_id;
+
+            if (array_key_exists($variant_id, $collectedVariants)) {
+                $line->quantity += $collectedVariants[$variant_id]['quantity'];
+                unset($collectedVariants[$variant_id]);
+            }
+
+            $line->save();
+        });
+
+        $basket->lines()->createMany($collectedVariants);
     }
 
     /**