+++ src/Core/Payments/Events/ThreeDSecureAttemptEvent.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace GetCandy\Api\Core\Payments\Events;
+
+use Illuminate\Broadcasting\Channel;
+use Illuminate\Queue\SerializesModels;
+use Illuminate\Broadcasting\PrivateChannel;
+use Illuminate\Foundation\Events\Dispatchable;
+use Illuminate\Broadcasting\InteractsWithSockets;
+
+class ThreeDSecureAttemptEvent
+{
+    use Dispatchable, InteractsWithSockets, SerializesModels;
+
+    public $contents;
+
+    /**
+     * Create a new event instance.
+     *
+     * @return void
+     */
+    public function __construct($contents)
+    {
+        $this->contents = $contents;
+    }
+
+    /**
+     * Get the channels the event should broadcast on.
+     *
+     * @return Channel|array
+     */
+    public function broadcastOn()
+    {
+        return new PrivateChannel('payments');
+    }
+}